Removed all the "_" prefixing parameter names and instead added an xml-level tag save="false" to implement non-savable parameters (at the ParamTree level).
Improved Button to use rich text, so that "_" in strings is interpreted as
subscript introduction and enabled word wrap.

xmlPage now automatically adds ":\n%1" to xmlButtons if they link to a 
menu and the space before "..." is non-breakable.

Added a new convenience function "filePath()" to Globals. This is a
dangerous trend, but I think in this case it is warranted.