Added tree() and treep() methods to ParamTree to access a subtree.
Most acqCCD variables are now per-camera, but I am not quite done
and bugs remain.

Introduced named lamps and shutters to connections.xml. Each camera
may have an associated lamp and/or shutter. Cameras may share lamps
and shutters.

VideoProg now uses first camera for its avoidCCD functionality.
This is not the final desired result.

Formalizing the distinction between i/o lines that are or are not
referred to in the code.

Outstanding bug: gui triangles are not properly colored.

Renamed "FrameOx" to "Frame:Ox", "TrigOx" to "Trigger:Ox", etc.
Also: "VideoLight" is now "Lamp:Video".

Camera order is now implicit from order in xml file. Camera partner
IDs now only have to be specified in one member of the pair. The
back link is automatic.