Done basic testing of independent camera settings in linux. It appears to generate correct timing.
Fixed bug in autoitems that caused new buttons to remain hidden.

Added "exclude" tag to <auto> tag, with "*id" special value to exclude
the ID of the tabbed page.

Cannot find what's wrong with guitriangle. It draws the triangle, but
ignores the brush. Could it be a Qt error?

Moved "CAMERAS" enum to enums.xml and defined "self" there.